requirements.txt
g4f
speechrecognition
pyaudio
pyttsx3
colorama
python-dotenv